backend/auth/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
passlib[bcrypt]
PyJWT
prometheus-client
